fix import, add missing error listener
